[shelly] fix NumberFormatException on Shelly2RelayStatus.timerDuration (#17269)
- If 'Auto OFF' mode is enabled on a Shelly Plus 1 device and set to a
  value with decimal places a status update throws an NumberFormatException.

  Changed timerDuration/timer_duration from Integer to Double
  fixes the problem.

+++ bundles/org.openhab.binding.shelly/src/main/java/org/openhab/binding/shelly/internal/api2/Shelly2ApiJsonDTO.java
@@ -865,7 +865,7 @@ public class Shelly2ApiJsonDTO {
         @SerializedName("timer_started_at")
         public Double timerStartetAt;
         @SerializedName("timer_duration")
-        public Integer timerDuration;
+        public Double timerDuration;
         public Double apower;
         public Double voltage;
         public Double current;
